requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.30.0
pydantic>=1.8.2
aiofiles>=0.8.0
python-dotenv>=0.19.0
pandas